+++ src/app/core/services/service-photo-api.service.ts
