+++ src/app/my-orders/components/sent-order-list-item/sent-order-list-item.component.spec.ts
@@ -2,7 +2,7 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { MasterReadonlyApiCacheService, ServicesApiCache } from '@app/core/services/cache';
-import { ServicesApiService } from '@app/service/services/services-api.service';
+import { ServicesApiService } from '@app/core/services/services-api.service';
 
 import { IonicModule } from '@ionic/angular';
 import { TranslateModule } from '@ngx-translate/core';
