+++ src/app/service/components/service-editor-page/service-editor-page.component.spec.ts
@@ -1,5 +1,7 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
 import { IonicModule } from '@ionic/angular';
+import { TranslateModule } from '@ngx-translate/core';
 
 import { ServiceEditorPageComponent } from './service-editor-page.component';
 
@@ -7,10 +9,10 @@ describe('ServiceEditorPageComponent', () => {
   let component: ServiceEditorPageComponent;
   let fixture: ComponentFixture<ServiceEditorPageComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [ ServiceEditorPageComponent ],
-      imports: [IonicModule.forRoot()]
+      declarations: [ServiceEditorPageComponent],
+      imports: [TranslateModule.forRoot(), IonicModule.forRoot(), RouterTestingModule],
     }).compileComponents();
 
     fixture = TestBed.createComponent(ServiceEditorPageComponent);
