+++ src/app/service/service.module.ts
@@ -22,17 +22,17 @@ import { ServicePublishStepSevenFormService } from '@app/service/forms/service-p
 import { ServicePublishStepSixFormService } from '@app/service/forms/service-publish-step-six-form.service';
 import { ServicePublishStepTwoFormService } from '@app/service/forms/service-publish-step-two-form.service';
 import { ServicePublishGuardService } from '@app/service/guards/service-publish-guard.service';
-import { PricesApiService } from '@app/service/services/prices-api.service';
-import { ServiceLocationApiService } from '@app/service/services/service-location-api.service';
-import { ServicePhotoApiService } from '@app/service/services/service-photo-api.service';
+import { PricesApiService } from '@app/core/services/prices-api.service';
+import { ServiceLocationApiService } from '@app/core/services/service-location-api.service';
+import { ServicePhotoApiService } from '@app/core/services/service-photo-api.service';
 import { ServicePublishAuthStateManagerService } from '@app/service/services/service-publish-auth-state-manager.service';
 import { ServicePublishDataHolderService } from '@app/service/services/service-publish-data-holder.service';
 import { ServicePublishDataPreparerService } from '@app/service/services/service-publish-data-preparer.service';
 import { ServicePublishService } from '@app/service/services/service-publish.service';
-import { ServiceScheduleApiService } from '@app/service/services/service-schedule-api.service';
+import { ServiceScheduleApiService } from '@app/core/services/service-schedule-api.service';
 import { ServiceStepsNavigationService } from '@app/service/services/service-steps-navigation.service';
-import { ServiceTagsApiService } from '@app/service/services/service-tags-api.service';
-import { ServicesApiService } from '@app/service/services/services-api.service';
+import { ServiceTagsApiService } from '@app/core/services/service-tags-api.service';
+import { ServicesApiService } from '@app/core/services/services-api.service';
 import { ChainManagerService } from '@app/service/services/steps-navigation-chain/chain-manager.service';
 import { StepFinalHandlerService } from '@app/service/services/steps-navigation-chain/step-final-handler.service';
 import { StepFiveHandlerService } from '@app/service/services/steps-navigation-chain/step-five-handler.service';
